Refactor valid companies port interface
