+++ presentation/controllers/cli.py
@@ -18,8 +18,7 @@ from domain.ports.scraper_company_data_port import ScraperCompanyDataPort
 from domain.ports.scraper_nsd_port import ScraperNsdPort
 from domain.ports.scraper_statements_raw_port import ScraperStatementRawPort
 from domain.ports.scraper_stock_quote_port import ScraperStockQuotePort
-from domain.ports.valid_companies_read_port import ValidCompaniesReadPort
-from domain.ports.valid_companies_write_port import ValidCompaniesWritePort
+from domain.ports.valid_companies_port import ValidCompaniesPort
 from domain.services.service_company_data import CompanyDataService
 from domain.services.service_nsd import NsdService
 from domain.services.service_ratios import RatiosService
@@ -61,8 +60,7 @@ class Cli:
         policy: NsdPolicyPort,
         uow_factory: UowFactoryPort,
         http_client: AffinityHttpClientPort,
-        valid_companies_read_port: ValidCompaniesReadPort,
-        valid_companies_write_port: ValidCompaniesWritePort,
+        valid_companies_port: ValidCompaniesPort,
     ) -> None:
         """Initialize the CLI with injected ports."""
         # Store injected dependencies for later composition
@@ -89,8 +87,7 @@ class Cli:
         self.uow_factory = uow_factory
 
         self.byte_formatter = ByteFormatter()
-        self.valid_companies_read_port = valid_companies_read_port
-        self.valid_companies_write_port = valid_companies_write_port
+        self.valid_companies_port = valid_companies_port
 
     def __call__(self) -> None:
         return self.run()
@@ -201,8 +198,7 @@ class Cli:
 
             uow_factory=self.uow_factory,
             worker_pool=self.worker_pool,
-            valid_companies_read_port=self.valid_companies_read_port,
-            valid_companies_write_port=self.valid_companies_write_port,
+            valid_companies_port=self.valid_companies_port,
         )
 
         # run the service
